Add files via upload
These are the .jpg files I used for this initial rough draft (I love cats), so I had to use cat photo. The code is using LWE encryption schema, but without noise. The noise I was unable to decrypt image correctly, that is what I will be working on amending in the remainder of this semester.